Moving translate package to be internal only.
After some discussion on the PR, in order to keep things simple and
easy to follow, encoding is moved to be internal/translate.
